Rename Java plugin loader to JVM plugin loader
This is to prevent confusion with Java Edition - Velocity is built on the JVM and this handles plugins loaded into the same JVM Velocity is running in.
